+++ src/main/java/org/bob/siungongsi/config/CorsConfigForDev.java
@@ -31,29 +31,3 @@ class CorsConfigForDev {
     };
   }
 }
-
-@Profile("prod")
-@Configuration
-class CorsConfigForProd {
-
-  private final CorsProperties corsProperties;
-
-  public CorsConfigForProd(CorsProperties corsProperties) {
-    this.corsProperties = corsProperties;
-  }
-
-  @Bean
-  public WebMvcConfigurer corsConfigurer() {
-    return new WebMvcConfigurer() {
-      @Override
-      public void addCorsMappings(CorsRegistry registry) {
-        registry
-            .addMapping("/**") // 모든 엔드포인트에 대해 CORS 허용
-            .allowedOrigins(corsProperties.allowedOrigins().toArray(new String[0])) // 허용할 도메인
-            .allowedMethods("*") // 모든 HTTP 메서드 허용 (GET, POST, PUT, DELETE 등)
-            .allowedHeaders("*") // 모든 요청 헤더 허용
-            .allowCredentials(false); // 쿠키 포함 허용 (true 시 allowedOrigins에 * 사용 불가)
-      }
-    };
-  }
-}
